fix float fann

+++ src/nnue/nnue_type.h
@@ -23,9 +23,9 @@
 //using flBiasType = int32_t;
 //using flWeightType = int32_t;
 
-using outType = int32_t;
-using accumulatorType = int32_t;
-using biasType = int32_t;
-using weightType = int32_t;
+using outType = float;
+using accumulatorType = float;
+using biasType = float;
+using weightType = float;
 
 #endif
